+++ e2e-spa/CHANGELOG.md
@@ -1,10 +1,22 @@
 # @scale-codec/e2e-spa
 
-## 0.0.1
+## 0.0.2
+
 ### Patch Changes
 
-- Updated dependencies [1904251]
-- Updated dependencies [07d0d6e]
-- Updated dependencies [b427748]
-  - @scale-codec/definition-compiler@0.3.0
-  - @scale-codec/definition-runtime@0.1.3
+-   Updated dependencies [26c9dd0]
+-   Updated dependencies [c78d861]
+-   Updated dependencies [3441cdb]
+-   Updated dependencies [334a28d]
+    -   @scale-codec/definition-compiler@0.4.0
+    -   @scale-codec/definition-runtime@0.4.0
+
+## 0.0.1
+
+### Patch Changes
+
+-   Updated dependencies [1904251]
+-   Updated dependencies [07d0d6e]
+-   Updated dependencies [b427748]
+    -   @scale-codec/definition-compiler@0.3.0
+    -   @scale-codec/definition-runtime@0.1.3